fix: cambiar tipo de dato de cantidad a NUMERIC para permitir decimales en detalle de ventas

+++ src/main/java/com/inventario/repository/Impl/VentaRepositoryImpl.java
@@ -127,7 +127,7 @@ public class VentaRepositoryImpl implements VentaRepository {
                             rs.getInt("id"),
                             rs.getInt("venta_id"),
                             p,
-                            rs.getInt("cantidad"),
+                            rs.getDouble("cantidad"),
                             rs.getDouble("precio_unitario")
                     ));
                 }
